requirements.txt
flask>=2.3.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
gunicorn>=21.0.0
